Preserve Uchi namespace
Rails automatically adds subdirectories of /app to the autoload paths,
with some exceptions like the ones for views or assets. This means we
don't have to reference stuff in app/models as `Model::Stuff`, which is
good.

It also means we lose the Uchi namespace for stuff in app/uchi, which is
bad. We'd really like all Uchi stuff to be under the Uchi namespace in
order to avoid name collisions with other gems or the main app.

By removing the uchi subdirectory from the autoload paths and adding it back
with a namespace, we can preserve the Uchi namespace. Thanks to
https://github.com/fxn/zeitwerk/issues/250 for pointing us in the right
direction.