Vendor flowbite-components
This should help the installation process a bit, reduces dependencies,
and makes it easier for us to extend and contribute back to
https://github.com/substancelab/flowbite-components.

In particular it means that we can build and ship a CSS file with all
the styles we need. With the components in a separate repository
Tailwind couldn't pick up the classes used in those components.