Vendor pagy
This is a vendorization of the pagy gem version 43.0.0.rc1 - the best
pagination gem out there.

pagy is MIT licensed, see
https://github.com/ddnexus/pagy/blob/d70e443872a5b18da4e482454b3c8b4a1c86cb6b/LICENSE.txt.

Vendoring it allows us to avoid a dependency, and also lets us move it
under the Uchi namespace, so we can be included into projects that
already depend on pagy without causing conflicts.