Remove stuff from Pagy that we don't need
We rely on the :offset paginator only, so we can remove the other
paginators and the optimist dependency.

The offset paginator works out of the box with Rails, ActiveRecord and
has UI support without the need for Javascript or clientside storage.
That makes it the best fit for our needs where we need to integrate into
various applications with minimal dependencies and assumptions.

We also don't need their helper methods for generating
Bulma/Bootstrap HTML as we have our own rendering for that.