Toggle dark mode based on system setting
It seems that Flowbite 4 switched the dark mode toggle to being
class-based instead of based on `@media (prefers-color-scheme: dark)`.

From https://flowbite.com/docs/customize/dark-mode/:

> Tailwind CSS has a dark mode functionality enabled and the components
> from Flowbite also support it by using the “class” option.

> If you install the Flowbite Plugin inside your Tailwind project then
> dark mode via the class method will work automatically as we do this
> for you by default.

However, the default in TailwindCSS 4 is to use the `media` strategy,
which means that the dark mode styles will be applied based on the
user's system preferences. See https://tailwindcss.com/docs/dark-mode

I can't figure out how to enable dark mode in Flowbite 4 that uses the
media-based strategy using CSS (with custom-variant, I suppose?). So
I've opted for a JavaScript-based solution that adds the `dark` class
to the `html` element when the user's system preference is dark.

Relying on Javascript for this is not ideal, but it seems to be the way
to go about it in Flowbite 4 for now.