Switch to another UX model
This new behavior closer resembles the native select dropdown in that...

- It opens when activated, not when focused
- The currently selected option is highlighted
- The dropdown closes when an option is selected or when clicking
  outside
- The id of the associated record change on committing the selection, ie
  by clicking an option or pressing Enter or tab

The big benefit of this approach is splitting up the input field so it
no longer serves multiple purposes of displaying the selected value and
accepting user input. This makes it much clearer what is going on for
the user.

This brings us more in line with the common Combobox UI pattern outlined
at https://www.w3.org/WAI/ARIA/apg/patterns/combobox/examples/combobox-autocomplete-both/.